+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 
 ### Added
+- GATT Service: Broadcast Audio Scamn Service Server and Client (BASS 1.0)
+
 ### Fixed
 ### Changed